Updated SDCard error checks
1. Check image tool errors
2. Check file existance
3. Sudo added for dd
4. Updated Readme

+++ yocto_altera_sdcard.sh
@@ -6,29 +6,45 @@
 yellow='\E[1;33m'
 NC='\033[0m'
 
-
 if [ $# -eq 0 ]
 then
-  lsblk
-  echo -e "${yellow}Please provide a partition:${NC}"
+  sudo lsblk
+  echo -e "${yellow}Please provide a drive name:${NC}"
   read prt
+  echo -e "${yellow}Please provide an image name:${NC}"
+  read img
 else
   prt=${1}
 fi
 
-echo -e "${yellow}Creating Yocto Rocko SD card for Altera-SoC...${NC}"
-if [ -f ../*sdimage-cyclone5* ]
+
+echo -e "${yellow}Creating Yocto image for Altera-SoC...${NC}"
+cd ..
+if [ -f *sdimage-cyclone5* ]
 then
-  rm ../*sdimage-cyclone5*
+  rm *sdimage-cyclone5*
 fi
 
-cd ..
-wic create sdimage-cyclone5-arria5 -e core-image-minimal
-umount /dev/${prt}1
-umount /dev/${prt}2
-mv sdimage-cyclone5-*.direct sdimage-cyclone5.direct
-sudo dd if=./sdimage-cyclone5.direct of=/dev/${prt}
-umount /dev/${prt}1
-umount /dev/${prt}2
-cd $OLDPWD
-echo -e "${yellow}SD card is ready!${NC}"
+wic create sdimage-cyclone5-arria5 -e ${img}
+retval=$?
+if [ ${retval} -eq 0 ]
+then
+  if [ -f *sdimage-cyclone5* ]
+  then
+    umount /dev/${prt}1
+    umount /dev/${prt}2
+    mv sdimage-cyclone5-*.direct sdimage-cyclone5.direct
+    echo -e "${yellow}Flashing SD card...${NC}"
+    sudo dd if=./sdimage-cyclone5.direct of=/dev/${prt}
+    umount /dev/${prt}1
+    umount /dev/${prt}2
+    cd $OLDPWD
+    echo -e "${yellow}Success: SD card is ready!${NC}"
+  else
+    cd $OLDPWD
+    echo -e "${yellow}Fail: No image file was found!${NC}"
+  fi
+else
+  cd $OLDPWD
+  echo -e "${yellow}Fail: Image creating tool failed!${NC}"
+fi
